+++ src/app/(auth)/sign-in/components/auth-sign-in-form.tsx
@@ -8,7 +8,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Form, FormControl, FormField, FormItem, FormLabel } from '@/components/ui/form';
-import { signIn } from '@/lib/apis/auth-server.api';
+import { signIn } from '@/apis/auth-server.api';
 import { UserSignInDTO } from '@/types/DTO/user.dto';
 import SITE_MAP from '@/constants/site-map.constant';
