+++ src/app/(auth)/sign-in/components/auth-sign-in-with-google-button.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { Button } from '@/components/ui/button';
-import { signInWithGoogle } from '@/lib/apis/auth-browser.api';
+import { signInWithGoogle } from '@/apis/auth-browser.api';
 
 const AuthSignInWithGoogleButton = () => {
   return (
     <Button
       onClick={signInWithGoogle}
       variant="secondary"
-      className="before:bg-google-logo flex gap-1 before:block before:h-5 before:w-5 before:bg-contain before:bg-center before:bg-no-repeat"
+      className="flex gap-1 before:block before:h-5 before:w-5 before:bg-google-logo before:bg-contain before:bg-center before:bg-no-repeat"
     >
       구글로 계속하기
     </Button>
