+++ src/app/(client)/meal/post/page.tsx
@@ -1,5 +1,5 @@
 import Contents from '@/app/(client)/meal/post/components/upload-image-section';
-import { getAiResponses } from '@/lib/apis/analysis-request.api';
+import { getAiResponses } from '@/apis/analysis-request.api';
 
 const MealPostPage = async (): Promise<JSX.Element> => {
   const data = await getAiResponses();
