+++ src/app/(client)/my-page/components/edit-profile.tsx
@@ -6,8 +6,8 @@ import { z } from 'zod';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { SupabaseBucket } from '@/types/supabase-bucket.type';
-import { getUser, updateUser } from '@/lib/apis/user.api';
-import { uploadImage } from '@/lib/apis/storage.api';
+import { getUser, updateUser } from '@/apis/user.api';
+import { uploadImage } from '@/apis/storage.api';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
