+++ src/app/(client)/my-page/page.tsx
@@ -1,6 +1,6 @@
 import CtaExampleFeedbackBanner from '@/components/commons/cta-example-feedback-banner';
 import GlassBackground from '@/components/commons/glass-background';
-import { getUser } from '@/lib/apis/user.api';
+import { getUser } from '@/apis/user.api';
 import ProfileBar from './components/profile-bar';
 import UserGoalCard from './components/user-goal-card';
 import UserPhysicalInfoCard from './components/user-physical-info-card';
