+++ src/app/api/meal-post/route.ts
@@ -1,4 +1,4 @@
-import { createFoodAnalysisRequestDetail } from '@/lib/apis/analysis-request.api';
+import { createFoodAnalysisRequestDetail } from '@/apis/analysis-request.api';
 import { NextResponse } from 'next/server';
 
 export async function POST(req: Request) {
