+++ src/store/user-store.ts
@@ -2,7 +2,7 @@ import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 import { UserDTO } from '@/types/DTO/user.dto';
 import { browserClient } from '@/lib/utils/supabase/client.util';
-import { getUser } from '@/lib/apis/user.api';
+import { getUser } from '@/apis/user.api';
 
 type UserStore = {
   user: UserDTO;
